Move assertAskQuery to extra module
