+++ tests/test_minimal_integration.py
@@ -1,13 +1,6 @@
 from rdflib import Graph
 from shaclgen.shaclgen import data_graph
-
-
-def assertAskQuery(graph, query):
-    result = graph.query(query)
-    if not result.askAnswer:
-        serialization = graph.serialize(format="turtle")
-        explanation = "The actual resulting shacl graph is:\n" + serialization
-    assert result.askAnswer, explanation
+from helpers import assertAskQuery
 
 
 def test_create_shape():
